Now do the ensemble
